act: remove shaded fork join pool implementation
I ran a series of benchmarks and there seem to be different behaviors between the existing and JDK
ForkJoinPools especially on JDK11 and Graal. The results have been really inconclusive about which one would be a benefit overall.

In lack of better evidence, there is not much reason to keep our own copy, so there it goes.

(I couldn't reproduce the high system CPU usage with more recent JDKs that Patrik noticed when
we had it on the table the last time in #22560.)

Fixes #26180.